+++ src/main/java/net/silentchaos512/scalinghealth/ScalingHealth.java
@@ -34,7 +34,7 @@ public class ScalingHealth implements IModBase {
     public static final String MOD_ID_OLD = "ScalingHealth";
     public static final String MOD_ID_LOWER = "scalinghealth";
     public static final String MOD_NAME = "Scaling Health";
-    public static final String VERSION = "1.3.36";
+    public static final String VERSION = "1.3.37";
     public static final String VERSION_SILENTLIB = "3.0.13";
     public static final int BUILD_NUM = 0;
     public static final String DEPENDENCIES = "required-after:forge@[14.23.5.2779,);required-after:silentlib@[" + VERSION_SILENTLIB + ",);after:morpheus";